+++ src/onegov.election_day/onegov/election_day/assets/js/common.js
@@ -3,3 +3,7 @@ $(document).foundation();
 
 // stack tables wishing to be stacked
 $('.stackable').stacktable();
+
+// force all dropdowns to be rendered in the direction specified in the
+// options of said dropdown (so if we say align:left, *always* align left)
+Foundation.libs.dropdown.small = function() { return false; };