+++ src/onegov.election_day/onegov/election_day/views/vote.py
@@ -1,4 +1,7 @@
+from morepath.request import Response
 from onegov.ballot import Ballot, Vote
+from onegov.core.csv import convert_list_of_dicts_to_csv
+from onegov.core.csv import convert_list_of_dicts_to_xlsx
 from onegov.core.security import Public
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.layout import DefaultLayout
@@ -15,6 +18,27 @@ def view_vote(self, request):
         'layout': layout,
         'counted': self.counted
     }
+
+
+@ElectionDayApp.json(model=Vote, name='json', permission=Public)
+def view_vote_as_json(self, request):
+    return self.export()
+
+
+@ElectionDayApp.view(model=Vote, name='csv', permission=Public)
+def view_vote_as_csv(self, request):
+    return convert_list_of_dicts_to_csv(self.export())
+
+
+@ElectionDayApp.view(model=Vote, name='xlsx', permission=Public)
+def view_vote_as_xlsx(self, request):
+    return Response(
+        convert_list_of_dicts_to_xlsx(self.export()),
+        content_type=(
+            'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
+        ),
+        content_disposition='inline; filename={}.xlsx'.format(self.title)
+    )
 
 
 @ElectionDayApp.json(model=Ballot, permission=Public, name='by-municipality')